DJANGO_PV311/DjangoAPI/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.1.1
requests==2.31.0 